rename intersectionMerge
